--- truckmanagement-transportation-microservice/src/main/java/com/example/truckmanagementtransportationmicroservice/TruckmanagementTransportationMicroserviceApplication.java
@@ -1,13 +1,0 @@
-package com.example.truckmanagementtransportationmicroservice;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class TruckmanagementTransportationMicroserviceApplication {
-
-    public static void main(String[] args) {
-        SpringApplication.run(TruckmanagementTransportationMicroserviceApplication.class, args);
-    }
-
-}
