--- truckmanagement-transportation-microservice/src/main/java/com/example/truckmanagementtransportationmicroservice/controller/Controller.java
@@ -1,5 +1,0 @@
-package com.example.truckmanagementtransportationmicroservice.controller;
-
-public class Controller {
-
-}
