--- truckmanagement-transportation-microservice/src/main/java/com/example/truckmanagementtransportationmicroservice/repo/Repo.java
@@ -1,5 +1,0 @@
-package com.example.truckmanagementtransportationmicroservice.repo;
-
-public class Repo {
-
-}
