--- truckmanagement-transportation-microservice/src/main/java/com/example/truckmanagementtransportationmicroservice/service/Services.java
@@ -1,5 +1,0 @@
-package com.example.truckmanagementtransportationmicroservice.service;
-
-public class Services {
-
-}
