--- truckmanagement-transportation-microservice/src/main/java/com/example/truckmanagementtransportationmicroservice/service/ServicesInterface.java
@@ -1,5 +1,0 @@
-package com.example.truckmanagementtransportationmicroservice.service;
-
-public interface ServicesInterface {
-
-}
