+++ truckmanagement-transportation-microservice/src/main/java/com/transportation/microservice/MicroserviceDriver.java
@@ -1,0 +1,11 @@
+package com.transportation.microservice;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class MicroserviceDriver {
+  public static void main(String[] args) {
+    SpringApplication.run(MicroserviceDriver.class, args);
+  }
+}
