+++ truckmanagement-transportation-microservice/src/main/java/com/transportation/microservice/controller/Controller.java
@@ -1,0 +1,6 @@
+package com.transportation.microservice.controller;
+
+
+public class Controller {
+
+}
