+++ truckmanagement-transportation-microservice/src/main/java/com/transportation/microservice/repo/RepoInterface.java
@@ -1,0 +1,5 @@
+package com.transportation.microservice.repo;
+
+public interface RepoInterface {
+
+}
