+++ truckmanagement-transportation-microservice/src/main/java/com/transportation/microservice/service/Services.java
@@ -1,0 +1,5 @@
+package com.transportation.microservice.service;
+
+public class Services {
+
+}
